services (#13)
* Add import for Service models

Include ServiceRequestModel and ServiceResponseModel to allow for operations related to service entities. This enhances the module's capabilities and ensures that service-related requests and responses are correctly handled.

* Add Service models for TCP/UDP with validation

Introduced `ServiceRequestModel` and `ServiceResponseModel` to represent service configurations. The models include validation rules for required fields and protocol specifications (TCP/UDP) with associated ports and timeout settings. This addition improves service handling in Palo Alto Networks' Strata Cloud Manager.

* Add Service class to SCM configuration objects

Introduced a new Service class to manage services in the Palo Alto Networks' Strata Cloud Manager. The class includes methods for creating, retrieving, updating, and listing services, with validation to ensure container parameters are correctly provided.

* Add 'tags' filter handling in Service and update tests

Extended the service.py to handle 'tags' filter in the 'list' method. Added comprehensive unit tests for services including the new 'tags' filter functionality and updated test cases for addresses and models to ensure proper validation.

* Add Service management and validation updates.

Refactor test validation to use correct exception classes. Document new Service capabilities in release notes and README. Increment version to 0.1.4 in pyproject.toml and add detailed service usage and model documentation.

* Add Service and Service Models to SDK documentation

Updated the SDK documentation index to include links to the new Service and Service Models sections. This ensures all parts of the SDK are properly documented and accessible.

* Add new Service documentation and update client usage

Added new entries in mkdocs.yml for Service and Service Models under SDK. Updated the index page description for better clarity and added a code example in the SDK client documentation for listing configured addresses.

* Add section for managing services in README

Updated the table of contents in README.md to include managing services. This addition covers listing services and creating a service, ensuring clarity and structure for users.

* Fix incorrect service label in mkdocs.yml

Correct the label for the service documentation from "Application" to "Service" to improve clarity and accuracy in the SDK documentation. This ensures that links and references within the documentation are properly labeled and user-friendly.