logging-switch (#15)
* Update logger level from DEBUG to INFO

This change adjusts the logging level from DEBUG to INFO in the logger setup. It applies to both the logger itself and the console handler, helping to reduce verbosity in log output.

* Update version to 0.1.6 and set default logging to INFO

This commit updates the project version from 0.1.5 to 0.1.6 in `pyproject.toml`. It also modifies the release notes to reflect the new version and changes the default logging level to INFO.