Handle empty API response content in SCM client (#19)
* Handle empty API response content in SCM client

Ensures the client does not break when the API returns empty or whitespace content. Modified the response handling in `scm/client.py` to return `None` instead of attempting to parse empty content as JSON.

* Add API client creation and extend rule attributes

This commit introduces the creation of an API client object using an example from the `scm.client` module. Additionally, it extends rule attributes in the WildfireAntivirusProfile configuration to include applications and file types.

* Add termynal for code examples and update release notes

Introduce termynal containers for improved code snippet clarity in the getting-started and troubleshooting docs. Also, document the bug fix in the new version 0.1.10 supporting empty API responses.

* Update version to 0.1.10

This update increments the version number from 0.1.9 to 0.1.10 in the pyproject.toml file. This change is necessary to reflect the latest modifications and improvements in the SDK.